feat: favorite add to playlist

+++ src/KanaPlayer/Converters/TimestampHumanizerConverter.cs
@@ -8,7 +8,7 @@ public class TimestampHumanizerConverter: IValueConverter
 {
     public object? Convert(object? value, Type targetType, object? parameter, CultureInfo culture) =>
         value is long timestamp
-            ? DateTimeOffset.FromUnixTimeSeconds((long)timestamp).DateTime.Date.ToString("yyyy-MM-dd")
+            ? DateTimeOffset.FromUnixTimeSeconds(timestamp).DateTime.Date.ToString("yyyy-MM-dd")
             : value;
 
     public object ConvertBack(object? value, Type targetType, object? parameter, CultureInfo culture) =>
